requirements.txt
streamlit>=1.23.0
torch>=2.0.0
numpy>=1.22.0
opencv-python>=4.7.0
Pillow>=9.5.0
pandas>=2.0.0
matplotlib>=3.7.0
av>=10.0.0
streamlit-webrtc>=0.47.0